Separated remaining tests
