+++ lib/baseClient.js
@@ -1,10 +1,8 @@
-"use strict";
-
-const urlLib = require('url');
-const headers = require('./headers.js');
+import * as urlLib from 'url'
+import {addUserAgent} from './headers.js'
 
 
-class BaseClient {
+export class BaseClient {
 	constructor(operation, params) {
 		this.operation = operation;
 		this.params = params;
@@ -59,14 +57,10 @@ class BaseClient {
 			}
 			this.options.headers['Content-Type'] = this.params.contentType || 'text/plain';
 		}
-		headers.addUserAgent(this.options.headers);
+		addUserAgent(this.options.headers);
 		if (this.params.secureProtocol) {
 			this.options.secureProtocol = this.params.secureProtocol;
 		}
 	}
-}
-
-module.exports = {
-	BaseClient,
 }
