+++ lib/httpClient.js
@@ -5,6 +5,8 @@ import * as querystring from 'querystring'
 import * as websocket from 'websocket'
 import {HighResolutionTimer} from './hrtimer.js'
 import {addUserAgent} from './headers.js'
+import * as agentkeepalive from 'agentkeepalive'
+import * as HttpsProxyAgent from 'https-proxy-agent'
 
 
 /**
@@ -44,7 +46,7 @@ class HttpClient {
 		}
 		this.options.agent = false;
 		if (this.params.agentKeepAlive) {
-			const KeepAlive = (this.options.protocol == 'https:') ? require('agentkeepalive').HttpsAgent : require('agentkeepalive');
+			const KeepAlive = (this.options.protocol == 'https:') ? agentkeepalive.HttpsAgent : agentkeepalive;
 			let maxSockets = 10;
 			if (this.params.requestsPerSecond) {
 				maxSockets += Math.floor(this.params.requestsPerSecond);
@@ -127,7 +129,6 @@ class HttpClient {
 		if (this.options.protocol == 'ws:') {
 			lib = websocket;
 		}
-		const HttpsProxyAgent = require('https-proxy-agent');
 
 		// adding proxy configuration
 		if (this.params.proxy) {
