+++ lib/loadtest.js
@@ -1,18 +1,9 @@
-'use strict';
-
-/**
- * Load Test a URL, website or websocket.
- * (C) 2013 Alex Fernández.
- */
-
-
-const http = require('http');
-const https = require('https');
-const testing = require('testing');
-const httpClient = require('./httpClient.js');
-const websocket = require('./websocket.js');
-const {Latency} = require('./latency.js');
-const {HighResolutionTimer} = require('./hrtimer.js');
+import * as http from 'http'
+import * as https from 'https'
+import * as httpClient from './httpClient.js'
+import * as websocket from './websocket.js'
+import {Latency} from './latency.js'
+import {HighResolutionTimer} from './hrtimer.js'
 
 const SHOW_INTERVAL_MS = 5000;
 
@@ -40,7 +31,7 @@ https.globalAgent.maxSockets = 1000;
  *	- quiet: do not log any messages (deprecated).
  * An optional callback will be called if/when the test finishes.
  */
-exports.loadTest = function(options, callback) {
+export function loadTest(options, callback) {
 	if (!options.url) {
 		console.error('Missing URL in options');
 		return;
@@ -63,7 +54,7 @@ exports.loadTest = function(options, callback) {
 	const operation = new Operation(options, callback);
 	operation.start();
 	return operation;
-};
+}
 
 /**
  * Used to keep track of individual load test Operation runs.
@@ -184,93 +175,5 @@ class Operation {
 			this.latency.show();
 		}
 	}
-}
-
-/**
- * A load test with max seconds.
- */
-function testMaxSeconds(callback) {
-	const options = {
-		url: 'http://localhost:7357/',
-		maxSeconds: 0.1,
-		concurrency: 1,
-	};
-	exports.loadTest(options, callback);
-}
-
-
-/**
- * A load test with max seconds.
- */
-function testWSEcho(callback) {
-	const options = {
-		url: 'ws://localhost:7357/',
-		maxSeconds: 0.1,
-		concurrency: 1,
-	};
-	exports.loadTest(options, callback);
-}
-
-function testIndexParam(callback) {
-	const options = {
-		url: 'http://localhost:7357/replace',
-		concurrency:1,
-		maxSeconds: 0.1,
-		indexParam: "replace"
-	};
-	exports.loadTest(options, callback);
-}
-
-function testIndexParamWithBody(callback) {
-	const options = {
-		url: 'http://localhost:7357/replace',
-		concurrency:1,
-		maxSeconds: 0.1,
-		indexParam: "replace",
-		body: '{"id": "replace"}'
-	};
-	exports.loadTest(options, callback);
-}
-
-function testIndexParamWithCallback(callback) {
-	const options = {
-		url: 'http://localhost:7357/replace',
-		concurrency:1,
-		maxSeconds: 0.1,
-		indexParam: "replace",
-		indexParamCallback: function() {
-			//https://gist.github.com/6174/6062387
-			return Math.random().toString(36).substring(2, 15) + Math.random().toString(36).substring(2, 15);
-		}
-	};
-	exports.loadTest(options, callback);
-}
-
-function testIndexParamWithCallbackAndBody(callback) {
-	const options = {
-		url: 'http://localhost:7357/replace',
-		concurrency:1,
-		maxSeconds: 0.1,
-		body: '{"id": "replace"}',
-		indexParam: "replace",
-		indexParamCallback: function() {
-			//https://gist.github.com/6174/6062387
-			return Math.random().toString(36).substring(2, 15) + Math.random().toString(36).substring(2, 15);
-		}
-	};
-	exports.loadTest(options, callback);
-}
-
-
-/**
- * Run all tests.
- */
-exports.test = function(callback) {
-	testing.run([testMaxSeconds, testWSEcho, testIndexParam, testIndexParamWithBody, testIndexParamWithCallback, testIndexParamWithCallbackAndBody], callback);
-};
-
-// run tests if invoked directly
-if (__filename == process.argv[1]) {
-	exports.test(testing.show);
 }
