+++ lib/websocket.js
@@ -1,14 +1,5 @@
-'use strict';
-
-/**
- * Load test a websocket.
- * (C) 2013 Alex Fernández.
- */
-
-
-const WebSocketClient = require('websocket').client;
-const testing = require('testing');
-const BaseClient = require('./baseClient.js').BaseClient;
+import * as websocket from 'websocket'
+import {BaseClient} from './baseClient.js'
 
 let latency;
 
@@ -16,9 +7,9 @@ let latency;
 /**
  * Create a client for a websocket.
  */
-exports.create = function(operation, params) {
+export function create(operation, params) {
 	return new WebsocketClient(operation, params);
-};
+}
 
 /**
  * A client that connects to a websocket.
@@ -36,7 +27,7 @@ class WebsocketClient extends BaseClient {
 	 * Start the websocket client.
 	 */
 	start() {
-		this.client = new WebSocketClient();
+		this.client = new websocket.client();
 		this.client.on('connectFailed', () => {});
 		this.client.on('connect', connection => this.connect(connection));
 		this.client.connect(this.params.url, []);
@@ -139,30 +130,5 @@ class WebsocketClient extends BaseClient {
 			}
 		}
 	}
-}
-
-
-function testWebsocketClient(callback) {
-	const options = {
-		url: 'ws://localhost:7357/',
-		maxSeconds: 0.1,
-		concurrency: 1,
-	};
-	exports.create({}, options);
-	testing.success(callback);
-}
-
-/**
- * Run tests, currently nothing.
- */
-exports.test = function(callback) {
-	testing.run([
-		testWebsocketClient,
-	], callback);
-};
-
-// start tests if invoked directly
-if (__filename == process.argv[1]) {
-	exports.test(testing.show);
 }
