+++ test/all.js
@@ -9,6 +9,9 @@ import {test as testHeaders} from './headers.js'
 import {test as testLatency} from './latency.js'
 import {test as testHttpClient} from './httpClient.js'
 import {test as testServer} from './testserver.js'
+import {test as testRequestGenerator} from './request-generator.js'
+import {test as testLoadtest} from './loadtest.js'
+import {test as testWebsocket} from './websocket.js'
 
 
 /**
@@ -17,7 +20,7 @@ import {test as testServer} from './testserver.js'
 function test() {
 	const tests = [
 		testHrtimer, testHeaders, testLatency, testHttpClient,
-		testServer,
+		testServer, testRequestGenerator, testLoadtest, testWebsocket,
 	];
 	testing.run(tests, 4200);
 }
